Fixed a NullPointerException that could occur when getting date elements. Also fixes SND missing links to the studies.
The NullPointerException occurred because getting the value of an attribute returns null if the attribute isn't present. An additional filter was added to exclude all null attributes before filtering further.

The missing SND links were caused by the repository handler throwing away all elements that didn't have a language set. The setting to assume a language has been enabled as a temp fix.

This commit also includes the prototypes for Prometheus metrics.

This resolves #142.