#217: Delete inactive records
Currently records marked as deleted in remote repositories are marked as inactive in the Elasticsearch index. SearchKit filters out these records by requiring records to have isActive set to true.

This commit changes this behaviour to delete records marked as inactive. This reduces the storage required for Elasticsearch, as well as allowing simplifications of SearchKit as it no longer needs to perform this filtering.