#205: Simplify the Elasticsearch index creation logic, rename FileHandler to ResourceHandler and make all the methods static
The prepareIndex and createIndex methods have been merged into a single createIndex method.
The fallback path creating an index with no settings or mappings has been removed, as the indices created in this fashion would cause errors due to the missing settings
Index settings are now always language dependent and mappings are the same for all indices.