#335: Fix nested fields not being searchable
Nested fields are not searchable by simple_query_string by default. For these fields to be searchable, they need to be copied to a search field in the root document.

This commit also fixes a bug that would cause an exception to be thrown if Elasticsearch didn't return a scroll id.