#146: Add support for indexing a repository from a specific folder.
A new configuration entry, path, has been added for each repository to allow a path to be specified. If configured, the directory is recursively searched and all files are parsed. The XML documents are only parsed once.

To avoid high memory usage the indexer uses a streaming model to parse each document to a CMMStudy independently.

The document response validator has been refactored to always use the JDom implementation, and the W3C implementation has been removed.

fixup! #146: Add support for indexing a repository from a specific folder.