#652: Add support for parsing DDI 3.x fragment documents
This commits adds a sample DDI 3.x fragment document and tests to validate the correct parsing of DDI 3.x fragment documents.

To enable support of parsing references, the XMLMapper was extracted to an interface. A new ResolvingXMLMapper was created to handle cases where elements can both be directly included in a StudyUnit and referenced.