#290: Add identifier information, such as an ORCID, to the creators field
This commit converts the "creators" field from a string to an object containing the name, affiliation and identifier as separate fields. The identifier is an object containing the identifier, its type and a link.

This is a breaking change to the CMMStudy model and will require reindexing the Elasticsearch indices when merged. The major version number has been amended accordingly.