Use the Spring PathMatchingResourcePatternResolver to load thematic indexing configuration, use ElasticsearchSet for searches
The method of using the Java resource mechanism to find directories in the filesystem will not work if the indexer is packaged into a Jar file. Using PathMatchingResourcePatternResolver is portable regardless of where the resources can be found.

Using ElasticsearchSet to perform searches reduces memory consumption as only part of the results need to be loaded at any one time. It also bypasses search size limitations with the use of the Elasticsearch Scroll API.

The error handling has been reworked to reduce the number of locations where errors are caught in order to make the code easier to understand.